+++ SnowRunner-Tool/Options.cs
@@ -5,7 +5,7 @@ namespace SnowRunner_Tool
 {
     public class Options
     {
-        [Option("platform", Required = true, HelpText = "Set your platform (Epic/Steam/Microsoft)")]
+        [Option("platform", Required = true, HelpText = "Set your platform (Epic/Steam)")]
         public string Platform { get; set; }
     }
 }